add tests for variable and version for major module version node, fix squashing algo
FossilOrigin-Name: 10e08bdf56ec846806309133ad9512b65c466018d334ffc2c631b82142112d5d

+++ parser/ix3/ast/meta.hpp
@@ -24,8 +24,8 @@ namespace meta {
 	bool operator == (const version& left, const version& right) noexcept ;
 
 	struct version : meta_parameter, x3::position_tagged {
-		std::int64_t major_v;
-		std::int64_t minor_v;
+		std::int32_t major_v;
+		std::int32_t minor_v;
 		~version() noexcept override =default ;
 		std::string_view name() const override ;
 		std::string value() const override ;
